fix(editor): rename editor.slash to .svelte.ts so runes compile
Svelte 5 only compiles runes ($state/$derived/etc.) in .svelte files and
.svelte.ts/.js files. The previous editor.slash.ts referenced $state(...)
inside buildSuggestionRender → at runtime that resolves to no global of
that name, throws ReferenceError, and the slash menu silently fails to
mount.

Renamed → editor.slash.svelte.ts so the Vite plugin transforms runes.
Updated three imports accordingly (Editor.svelte, editor.extensions.ts,
index.ts). No API changes.
