+++ src/lib/Editor/editor.slash.svelte.ts
