+++ src/lib/Editor/index.ts
@@ -1,5 +1,5 @@
 export { default as Editor } from './Editor.svelte'
-export { buildDefaultSlashCommands } from './editor.slash.js'
+export { buildDefaultSlashCommands } from './editor.slash.svelte.js'
 export type {
     EditorProps,
     EditorApi,
